Addev add event modal and styled a bit

+++ WebApplicationVentixe/Models/Events/AddEventFormData.cs
@@ -10,7 +10,7 @@ public class AddEventFormData
     public DateTime EventDate { get; set; }
 
     public int SilverTicketAmount { get; set; }
-    public int SilverTicketPrice { get; set; }
+    public double SilverTicketPrice { get; set; }
     public int GoldTicketAmount { get; set; }
-    public int GoldTicketPrice { get; set; }
+    public double GoldTicketPrice { get; set; }
 }
